scripts/codegen-parser/requirements.txt
pytest==7.4.4
PyYAML==6.0.1
